refactor containers into more organized folders
